# lib/api-zod/src/generated/types/communication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NGO Donor Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Communication {
  id: string;
  /** @nullable */
  donorId?: string | null;
  subject: string;
  body: string;
  type: string;
  /** @nullable */
  sentAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}