# lib/api-zod/src/generated/types/createBeneficiaryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NGO Donor Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBeneficiaryBodyHome } from "./createBeneficiaryBodyHome";
import type { CreateBeneficiaryBodyStatus } from "./createBeneficiaryBodyStatus";

export interface CreateBeneficiaryBody {
  name: string;
  /** @nullable */
  dateOfBirth?: Date | null;
  /** @nullable */
  gender?: string | null;
  home: CreateBeneficiaryBodyHome;
  status: CreateBeneficiaryBodyStatus;
  /** @nullable */
  medicalInfo?: string | null;
  /** @nullable */
  educationInfo?: string | null;
  /** @nullable */
  notes?: string | null;
}