# lib/api-zod/src/generated/types/createDonationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NGO Donor Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDonationBody {
  donorId: string;
  amount: number;
  currency: string;
  donationDate: Date;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  receiptNumber?: string | null;
  /** @nullable */
  notes?: string | null;
}